+++ src/main/java/com/ssafy/rualone/domain/attraction/entity/Gugun.java
@@ -1,0 +1,4 @@
+package com.ssafy.rualone.domain.attraction.entity;
+
+public class Gugun {
+}
