+++ src/main/java/com/ssafy/rualone/domain/board/application/PostServiceImpl.java
@@ -1,6 +1,6 @@
 package com.ssafy.rualone.domain.board.application;
 
-import com.ssafy.rualone.domain.board.Post;
+import com.ssafy.rualone.domain.board.entity.Post;
 import com.ssafy.rualone.domain.board.dao.PostMapper;
 import com.ssafy.rualone.domain.board.dto.request.PostCreateRequest;
 import com.ssafy.rualone.domain.board.dto.request.PostUpdateRequest;
@@ -10,7 +10,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @Slf4j
