+++ src/main/java/com/ssafy/rualone/domain/board/dao/PostMapper.java
@@ -1,6 +1,6 @@
 package com.ssafy.rualone.domain.board.dao;
 
-import com.ssafy.rualone.domain.board.Post;
+import com.ssafy.rualone.domain.board.entity.Post;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
