+++ src/main/java/com/ssafy/rualone/domain/board/dto/request/PostCreateRequest.java
@@ -1,6 +1,6 @@
 package com.ssafy.rualone.domain.board.dto.request;
 
-import com.ssafy.rualone.domain.board.Post;
+import com.ssafy.rualone.domain.board.entity.Post;
 import lombok.Getter;
 import lombok.ToString;
